+++ src/Core/Config.php
@@ -95,6 +95,12 @@ class Config
      */
     protected $countryList = null;
 
+    /**
+     * is AmazonPayExpress as Paymentmethod active
+     *
+     * @var bool|null
+     */
+    protected $bIsAmazonExpressActive = null;
 
     /**
      * Checks if module configuration is valid
@@ -281,14 +287,6 @@ class Config
     /**
      * @return bool
      */
-    public function displayExpressInPDP(): bool
-    {
-        return (bool)Registry::getConfig()->getConfigParam('blAmazonPayExpressPDP');
-    }
-
-    /**
-     * @return bool
-     */
     public function useExclusion(): bool
     {
         return (bool)Registry::getConfig()->getConfigParam('blAmazonPayUseExclusion');
@@ -319,11 +317,42 @@ class Config
     }
 
     /**
+     * @param bool $bIsAdmin
      * @return bool
+     * @SuppressWarnings(PHPMD.BooleanArgumentFlag)
      */
-    public function displayExpressInMiniCartAndModal(): bool
+    public function displayExpressInPDP(bool $bIsAdmin = false): bool
     {
-        return (bool)Registry::getConfig()->getConfigParam('blAmazonPayExpressMinicartAndModal');
+        return $this->displayExpressButton('blAmazonPayExpressPDP', $bIsAdmin);
+    }
+
+    /**
+     * @param bool $bIsAdmin
+     * @return bool
+     * @SuppressWarnings(PHPMD.BooleanArgumentFlag)
+     */
+    public function displayExpressInMiniCartAndModal(bool $bIsAdmin = false): bool
+    {
+        return $this->displayExpressButton('blAmazonPayExpressMinicartAndModal', $bIsAdmin);
+    }
+
+    /**
+     * @param string $sVarName
+     * @param bool $bIsAdmin
+     * @return bool
+     * @SuppressWarnings(PHPMD.BooleanArgumentFlag)
+     */
+    private function displayExpressButton(string $sVarName, bool $bIsAdmin = false): bool
+    {
+        $bShowButton = (bool)Registry::getConfig()->getConfigParam($sVarName);
+        if (is_null($this->bIsAmazonExpressActive)) {
+            $this->bIsAmazonExpressActive = false;
+            $oPayment = oxNew(Payment::class);
+            $oPayment->load(Constants::PAYMENT_ID_EXPRESS);
+            $this->bIsAmazonExpressActive = $oPayment->isLoaded() &&
+                $oPayment->getFieldData('oxactive');
+        }
+        return $bShowButton && ($this->bIsAmazonExpressActive || $bIsAdmin);
     }
 
     /**